Update build_tree to add line to proper parent
Update logic for adding child line to siblings:
Push onto the front of the list instead of appending to the end.
Check the proper line property on the latest tree line when looking for the parent line.
Update build_tree to reverse all children, needed with change to sort of child lines.
Add tests.